requirements.txt
langchain_classic
langchain-openai>=0.1.6
python-dotenv
